+++ src/main/java/chess/presentation/view/PieceShapeSelector.java
@@ -1,4 +1,4 @@
-package chess.view;
+package chess.presentation.view;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
